requirements.txt
google-api-python-client==2.97.0
openai==1.9.0
python-dotenv==1.0.0
requests==2.31.0
